DreamBooth/sd_dreambooth_extension/requirements.txt
accelerate>=0.21.0
dadaptation>=3.1
diffusers>=0.23.1
discord-webhook==1.3.0
fastapi
gitpython>=3.1.40
pytorch_optimizer==2.12.0
Pillow
tqdm
tomesd>=0.1.2
transformers>=4.30.2